@@ -15,61 +17,65 @@
     "format": "prettier . --write"
   },
   "dependencies": {
-    "@headlessui/vue": "^1.7.2",
-    "@heroicons/vue": "^2.0.11",
-    "@popperjs/core": "^2.11.6",
+    "@headlessui/vue": "^1.7.13",
+    "@heroicons/vue": "^2.0.17",
+    "@popperjs/core": "^2.11.7",
     "body-scroll-lock": "^4.0.0-beta.0",
-    "pinia": "^2.0.22",
-    "vue": "^3.2.39",
-    "vue-router": "^4.1.5"
+    "pinia": "^2.0.35",
+    "vue": "^3.2.47",
+    "vue-router": "^4.1.6"
   },
   "devDependencies": {
-    "@babel/core": "^7.19.1",
-    "@babel/preset-env": "^7.19.1",
-    "@babel/types": "^7.19.0",
-    "@cypress/code-coverage": "^3.10.0",
-    "@cypress/vite-dev-server": "^3.1.1",
-    "@cypress/vue": "^4.2.0",
-    "@pinia/testing": "^0.0.14",
+    "@babel/core": "^7.21.4",
+    "@babel/preset-env": "^7.21.4",
+    "@babel/types": "^7.21.4",
+    "@badeball/cypress-cucumber-preprocessor": "^16.0.3",
+    "@bahmutov/cypress-esbuild-preprocessor": "^2.2.0",
+    "@cypress/browserify-preprocessor": "^3.0.2",
+    "@cypress/code-coverage": "^3.10.4",
+    "@cypress/vite-dev-server": "^5.0.5",
+    "@cypress/vue": "^5.0.5",
+    "@pinia/testing": "^0.0.16",
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "@tailwindcss/forms": "^0.5.3",
-    "@tailwindcss/typography": "^0.5.7",
-    "@types/cypress-cucumber-preprocessor": "^4.0.1",
-    "@types/node": "^18.7.18",
-    "@typescript-eslint/eslint-plugin": "^5.38.0",
-    "@typescript-eslint/parser": "^5.38.0",
+    "@tailwindcss/typography": "^0.5.9",
+    "@types/node": "^18.15.13",
+    "@typescript-eslint/eslint-plugin": "^5.59.0",
+    "@typescript-eslint/parser": "^5.59.0",
     "@vicons/fa": "^0.12.0",
-    "@vitejs/plugin-vue": "^3.1.0",
-    "@vitest/coverage-c8": "^0.23.4",
-    "@vitest/ui": "^0.23.4",
-    "@vue/compiler-sfc": "^3.2.39",
-    "@vue/test-utils": "^2.0.2",
-    "autoprefixer": "^10.4.12",
-    "babel-loader": "^8.2.5",
-    "c8": "^7.12.0",
-    "concurrently": "^7.4.0",
-    "cypress": "^10.8.0",
-    "cypress-cucumber-preprocessor": "^4.3.1",
-    "eslint": "^8.23.1",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-vue": "^9.5.1",
-    "happy-dom": "^6.0.4",
-    "naive-ui": "^2.33.3",
-    "postcss": "^8.4.16",
-    "postcss-import": "^15.0.0",
-    "postcss-scss": "^4.0.5",
-    "prettier": "^2.7.1",
-    "sass": "^1.55.0",
-    "stylelint": "^14.12.1",
-    "stylelint-config-recommended": "^9.0.0",
-    "tailwindcss": "^3.1.8",
-    "typescript": "^4.8.3",
+    "@vitejs/plugin-vue": "^4.1.0",
+    "@vitest/coverage-c8": "^0.30.1",
+    "@vitest/ui": "^0.30.1",
+    "@vue/compiler-dom": "^3.2.47",
+    "@vue/compiler-sfc": "^3.2.47",
+    "@vue/server-renderer": "^3.2.47",
+    "@vue/test-utils": "^2.3.2",
+    "autoprefixer": "^10.4.14",
+    "babel-loader": "^8.3.0",
+    "c8": "^7.13.0",
+    "concurrently": "^8.0.1",
+    "cypress": "^12.10.0",
+    "esbuild": "^0.17.18",
+    "eslint": "^8.39.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-vue": "^9.11.0",
+    "happy-dom": "^9.9.2",
+    "naive-ui": "^2.34.3",
+    "postcss": "^8.4.23",
+    "postcss-import": "^15.1.0",
+    "postcss-scss": "^4.0.6",
+    "prettier": "^2.8.7",
+    "sass": "^1.62.0",
+    "stylelint": "^15.5.0",
+    "stylelint-config-recommended": "^12.0.0",
+    "tailwindcss": "^3.3.1",
+    "typescript": "^5.0.4",
     "vfonts": "^0.0.3",
-    "vite": "^3.1.3",
-    "vite-plugin-istanbul": "^3.0.1",
-    "vite-tsconfig-paths": "^3.5.0",
-    "vitest": "^0.23.4",
+    "vite": "^4.3.1",
+    "vite-plugin-istanbul": "^4.0.1",
+    "vite-tsconfig-paths": "^3.6.0",
+    "vitest": "^0.30.1",
     "vue-tsc": "^0.40.13",
-    "webpack": "^5.76.0"
+    "webpack": "^5.80.0"
   }
 }
